feat(web): move search options into a modal
