+++ web/src/lib/components/shared-components/full-screen-modal.svelte
@@ -69,7 +69,7 @@
     </div>
     {#if isStickyBottom}
       <div
-        class="flex flex-col sm:flex-row justify-end w-full gap-2 sm:gap-4 sticky bottom-0 py-4 px-5 bg-immich-bg dark:bg-immich-dark-gray border-t border-gray-200 dark:border-gray-500 shadow"
+        class="flex flex-col sm:flex-row justify-end w-full gap-2 sm:gap-4 sticky bottom-0 py-4 px-5 bg-immich-bg dark:bg-immich-dark-gray border-t border-gray-200 dark:border-gray-500 shadow z-[9999]"
       >
         <slot name="sticky-bottom" />
       </div>
