+++ web/src/lib/i18n/en.json
@@ -1067,6 +1067,7 @@
   "search_for_existing_person": "Search for existing person",
   "search_no_people": "No people",
   "search_no_people_named": "No people named \"{name}\"",
+  "search_options": "Search options",
   "search_people": "Search people",
   "search_places": "Search places",
   "search_state": "Search state...",
